Get make_url_query to conftest + try to fix imports.
